requirements.txt
slack_sdk==3.33.3          # For Slack integration
requests==2.32.3           # For making HTTP requests to APIs (like CircleCI, Google Chat)
google-api-python-client==2.151.0  # For Google APIs (Gmail, Google Chat)
py-trello==0.20.1    # For Trello API integration
websocket-client==1.8.0    # Required for real-time communication in Slack
nltk==3.9.1                # For natural language processing in email summarization
python-dotenv==1.0.0       # To manage environment variables